% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{checkifnull}
\alias{checkifnull}
\title{check if null}
\usage{
checkifnull(input)
}
\arguments{
\item{input}{input variable. String or NULL}
}
\value{
If input is NULL, returns NULL. If input is string, returns input as symbol
(like "get(input)", but better for ggplot)
}
\description{
Checks if input variable is null. If not, returns variable as symbol.
This is needed for the evaluation within ggplot2, where NULL can be used as placeholders.
}
\examples{
checkifnull(NULL)
}
