% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gene_lolliplot.R
\name{create_gene_lolliplot}
\alias{create_gene_lolliplot}
\title{Create lolliplot gene plot}
\usage{
create_gene_lolliplot(
  gene_name = GOI,
  variant_filename = varFile,
  variant_input_type = varInputType,
  varPalette = varPalette,
  color_by_data = "published",
  exon_height = exonHeight,
  min_intron_size = minIntronSize
)
}
\arguments{
\item{gene_name}{\code{character} vector of the HGNC symbol gene of interest}

\item{variant_filename}{provided \code{csv} formatted variant file or \code{dataframe} of variants}

\item{variant_input_type}{\code{file} or \code{dataframe} depending on input provided}

\item{exon_height}{\code{numeric} representing the draw height of \code{exon} features}
}
\value{
output to gene symbol prefixed \code{SVG} formatted plots, and printed to local graphics device
}
\description{
Create lolliplot gene plot
}
