% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_peptide_lolliplot.R
\name{create_peptide_lolliplot}
\alias{create_peptide_lolliplot}
\title{Create lolliplot peptide plot}
\usage{
create_peptide_lolliplot(
  gene_name = GOI,
  variant_filename = varFile,
  variant_input_type = varInputType,
  varPalette = varPalette,
  color_by_data = "published",
  exon_height = exonHeight
)
}
\arguments{
\item{gene_name}{`character` vector of the HGNC symbol gene of interest}

\item{variant_filename}{provided `csv` formatted variant file or `dataframe` of variants}

\item{variant_input_type}{`file` or `dataframe` depending on input provided}

\item{exon_height}{`numeric` representing the draw height of `exon` features}
}
\value{
output to gene symbol prefixed `SVG` formatted plots, and printed to local graphics device
}
\description{
Create lolliplot peptide plot
}
