% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinvar.R
\name{get_clinvar}
\alias{get_clinvar}
\title{Get ClinVar variants from a specified gene via API}
\usage{
get_clinvar(gene)
}
\arguments{
\item{gene}{The gene from which the ClinVar variants should be retreived, e.g. "MBOAT7"}
}
\value{
A tibble containing the ClinVar variants with the transcript, gene name,
c_code, p_code and clinical significance
}
\description{
Takes a gene name as an input and returns the clinvar variants, retrieved via API
}
\examples{
MBOAT7_clinvar_variants = get_clinvar("MBOAT7", min_clinsig = 2)
}
