% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionickomics_utils.r
\name{get_EnsemblIDs}
\alias{get_EnsemblIDs}
\title{Get EnsemblID and Canonical transcript ID}
\usage{
get_EnsemblIDs(gene_of_interest = GOI)
}
\arguments{
\item{gene_of_interest}{`character` vector of the HGNC symbol of the gene of interest}
}
\value{
`transcript` and `exon` models
}
\description{
Get EnsemblID and Canonical transcript ID
}
