% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionickomics_utils.r
\name{get_gene_variants}
\alias{get_gene_variants}
\title{Get variants from NIH Clinical Tables}
\usage{
get_gene_variants(gene_name)
}
\arguments{
\item{gene_name}{character vector of HGNC symbol gene name}
}
\value{
dataframe of cleaned and filtered disease causing variants for the gene of interest
}
\description{
This function will obtain the latest information for a \code{gene_name}
}
\examples{
gene_gene_vaariants("CRB1")
}
