% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gnomad.R
\name{get_gnomad}
\alias{get_gnomad}
\title{get gnomAD variants per gene}
\usage{
get_gnomad(gene, account_outliers = T, outlier_threshold = 2000)
}
\arguments{
\item{gene}{The gene name, e.g. "MBOAT7"}

\item{account_outliers}{Boolean. Should we account for outliers, i.e. variants with an extraordinary high allele count?}

\item{outlier_threshold}{The threshold for the outliers, allele counts larger than this value will be collapsed (divided by 10)}
}
\value{
A tibble which contains the gnomad missense variants of the provided gene, with duplicated rows based on the allele count.
This will be needed for the histogram plot
}
\description{
Takes a gene name as an input and returns the gnomad missense variants, retrieved via API
}
\examples{
LRP6_gnomad = get_gnomad(gene = "LRP6", account_outliers = F)
}
