% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionickomics_utils.r
\name{get_variants_ClinVar}
\alias{get_variants_ClinVar}
\title{Title}
\usage{
get_variants_ClinVar(gene_of_interest = GOI, pathogenic = TRUE)
}
\arguments{
\item{pathogenic}{}
}
\description{
Title
}
