% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LRP6_cohort}
\alias{LRP6_cohort}
\title{LRP6 cohort data}
\format{
A dataframe
\describe{
\item{Individual}{Individual ID}
\item{variant_c}{c.Code of the variant}
\item{variant_p}{p.Code of the variant}
\item{origin}{origin, e.g. de novo or inherited}
\item{origin_simple}{simplified origin, i.e. maternal=inherited}
\item{classification}{classification of the variant}
\item{Cohort variant occurence}{How often does this variant occur in the cohort}
\item{protein_position}{amino acid position}
}
}
\usage{
LRP6_cohort
}
\description{
Imported cohort data with \code{\link{import_cohort}}
}
\keyword{datasets}
