% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MBOAT7_clinvar}
\alias{MBOAT7_clinvar}
\title{MBOAT7 clinvar data}
\format{
A dataframe
\describe{
\item{transcript}{Transcript}
\item{gene}{Gene}
\item{c_code}{c.Code of the variant}
\item{p_code}{p.Code of the variant}
\item{clinsig}{classification of the variant}
\item{protein_position}{amino acid position}
\item{clinsig_numeric}{classification of the variant encoded as numeric}
\item{clinsig_simple}{classification of the variant collapsed}
}
}
\usage{
MBOAT7_clinvar
}
\description{
Got clinvar data with \code{\link{get_clinvar}}
}
\keyword{datasets}
