% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MBOAT7_cohort}
\alias{MBOAT7_cohort}
\title{MBOAT7 cohort data}
\format{
A dataframe
\describe{
\item{Individual}{Individual ID}
\item{Family}{Family ID}
\item{Sex}{Sex of the individual}
\item{Age}{Age of the individual}
\item{Zygosity}{Zygosity of the variant}
\item{Variant_c}{c.Code of the variant}
\item{Variant_p}{p.Code of the variant}
\item{Consequence}{molecular consequence, missense or frameshift etc}
\item{Classification}{classification of the variant}
\item{LoF}{Loss of function variant? True or False}
\item{Cohort variant occurence}{How often does this variant occur in the cohort}
\item{protein_position}{amino acid position}
}
}
\usage{
MBOAT7_cohort
}
\description{
Imported cohort data with \code{\link{import_cohort}}
}
\keyword{datasets}
