% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MBOAT7_domains}
\alias{MBOAT7_domains}
\title{MBOAT7 domain data}
\format{
A dataframe
\describe{
\item{accession}{Pfam accession ID}
\item{name}{domain name}
\item{name_short}{short domain name}
\item{start}{start coordinates}
\item{end}{end coordinates}
\item{name_short_numbered}{unique name for each domain}
}
}
\usage{
MBOAT7_domains
}
\description{
Got domain data with \code{\link{get_domains}}
}
\keyword{datasets}
