% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MBOAT7_gnomad}
\alias{MBOAT7_gnomad}
\title{MBOAT7 clinvar data}
\format{
A dataframe
\describe{
\item{variant_id}{Variant ID from gnomad}
\item{consequence}{molecular consequence, missense or frameshift etc}
\item{homozygote_count}{number of homozygotes}
\item{c_code}{c.Code of the variant}
\item{p_code}{p.Code of the variant}
\item{protein_position}{amino acid position}
}
}
\usage{
MBOAT7_gnomad
}
\description{
Got gnomad data with \code{\link{get_gnomad}}
}
\keyword{datasets}
