% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lolliplot.R
\name{plot_lolliplot}
\alias{plot_lolliplot}
\title{Plot Lolliplot}
\usage{
plot_lolliplot(
  cohort = NULL,
  domains = NULL,
  clinvar = NULL,
  gnomad = NULL,
  density_adjustment = 500,
  color_domains_by = "accession",
  shorten_domainname = T,
  .y = "Cohort variant occurence",
  .label = NULL,
  .linetype = NULL
)
}
\arguments{
\item{cohort}{The cohort dataframe (see \code{\link{import_cohort}} for details)}

\item{domains}{The domain data (see \code{\link{get_domains}} for details)}

\item{clinvar}{The clinvar data (see \code{\link{get_clinvar}} for details)}

\item{gnomad}{The gnomad data (see \code{\link{get_gnomad}} for details)}

\item{density_adjustment}{Enlarge or shrink the gnomad density plot, multiplied by this numeric value}

\item{color_domains_by}{How the domains should be coloured.
One of "accession" (default), which colors same repeated domains in the same colour
(e.g. all LDLa domains, but not LDLb domains), "type", which colours same types like repeats or receptors,
or "unique", where all domains will get unique colours}

\item{shorten_domainname}{Boolean. Should the domain name be shortened?
This will strip the text after the first underscore, e.g. "LDL_recept_a" will be "LDL"}

\item{.y}{The cohort column name for the y aesthetic. Defaults to "Cohort variant occurence"}

\item{.label}{The cohort column name for the label aesthetic. Example: "variant_p"}

\item{.linetype}{The cohort column name for the linetype aesthetic. Example: "origin_simple"}
}
\value{
A ggplot object (the Lolliplot)
}
\description{
Plots a Lolliplot and takes the cohort data from \code{\link{import_cohort}},
clinvar data from \code{\link{get_clinvar}}, gnomad data from \code{\link{get_gnomad}}
and the domain data from \code{\link{get_domains}}
}
\examples{
LRP6_lolliplot = plot_lolliplot(cohort = LRP6_cohort, domains = LRP6_domains,
clinvar = LRP6_clinvar, gnomad = LRP6_gnomad, .linetype = "origin_simple")
}
