% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionickomics_utils.r
\name{process_HuVarOutput}
\alias{process_HuVarOutput}
\title{Process variants from HuVar output}
\usage{
process_HuVarOutput(filename = huvarfile, phenotypes_to_keep = "none")
}
\arguments{
\item{filename}{path to the HuVar output saved data}

\item{phenotypes_to_keep}{list object of characters detailing the disease phenotypes to be kept}
}
\value{
dataframe of cleaned and filtered HuVar single nucleotide variants
}
\description{
Process variants from HuVar output
}
