% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionickomics_utils.r
\name{read_variant_data}
\alias{read_variant_data}
\title{Read variant data}
\usage{
read_variant_data(varFile, do_count = FALSE)
}
\arguments{
\item{varFile}{\code{path} to local variant file}

\item{do_count}{default of \code{FALSE}, function to count the frequency of observations of each variant}
}
\value{
variant \code{dataframe}
}
\description{
Read variant data
}
\examples{
read_variant_data("variant.txt")
}
