% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionickomics_utils.r
\name{rescale_introns}
\alias{rescale_introns}
\title{Rescale introns}
\usage{
rescale_introns(grange, minIntronSize)
}
\arguments{
\item{grange}{`GRanges` object of the transcript model for the gene of interest}

\item{minIntronSize}{`numeric` vector to fix all `introns` to for representation}
}
\value{
rescaled `GRanges` object with updated `intron` sizes
}
\description{
Rescale introns
}
