% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionickomics_utils.r
\name{tidy_ClinVar}
\alias{tidy_ClinVar}
\title{Tidy ClinVar data}
\usage{
tidy_ClinVar(df)
}
\arguments{
\item{df}{}
}
\value{
processed
}
\description{
Tidy ClinVar data
}
\examples{
tidy_ClinVar(exampledata)
}
