% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ukl_colors}
\alias{ukl_colors}
\title{UKL colors}
\usage{
ukl_colors(n)
}
\arguments{
\item{n}{number of colors to return}
}
\value{
list of n colors
}
\description{
Provides a color palette based on the UKL corporate identity
}
\examples{
ukl_colors(3)
}
