% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionickomics_utils.r
\name{get_ClinVar_variants}
\alias{get_ClinVar_variants}
\title{Get ClinVar Variants}
\usage{
get_ClinVar_variants(gene_of_interest)
}
\arguments{
\item{gene_of_interest}{}
}
\value{
dataframe of tidied NCBI ClinVar variants
}
\description{
Get ClinVar Variants
}
