% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_domains.R
\name{get_domains}
\alias{get_domains}
\title{Get the pfam domain coordinates and descriptions}
\usage{
get_domains(gene)
}
\arguments{
\item{gene}{the name of the gene, e.g. BRCA1}
}
\value{
Domains
}
\description{
Get the pfam domain coordinates and descriptions
}
\examples{
domains = get_domains("LRP6")

}
