% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_domains.R
\name{get_domains}
\alias{get_domains}
\title{Get the pfam domain coordinates and descriptions}
\usage{
get_domains(gene)
}
\arguments{
\item{gene}{The name of the gene, e.g. "LRP6"}
}
\value{
A tibble containing the start/end positions and domain names of the specified gene
}
\description{
Takes a gene name as an input and returns the Pfam domain coordinates, retrieved via API
}
\examples{
LRP6_domains = get_domains("LRP6")
}
