% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gnomad.R
\name{get_gnomad}
\alias{get_gnomad}
\title{Get gnomAD variants per gene}
\usage{
get_gnomad(
  gene,
  account_outliers = TRUE,
  outlier_threshold = 2000,
  outlier_shrinkfactor = 10
)
}
\arguments{
\item{gene}{The gene name, e.g. "KDM2A"}

\item{account_outliers}{Boolean. Should we account for outliers, i.e. variants with an extraordinary high allele count?}

\item{outlier_threshold}{The threshold for the outliers, allele counts larger than this value will be collapsed (divided by \code{outlier_shrinkfactor})}

\item{outlier_shrinkfactor}{Allele counts larger than \code{outlier_threshold} will be divided by this value}
}
\value{
A tibble which contains the gnomad missense variants of the provided gene, with duplicated rows based on the allele count.
This will be needed for the histogram plot
}
\description{
Takes a gene name as an input and returns the gnomad missense variants, retrieved via API
}
\examples{
KDM2A_gnomad = get_gnomad(gene = "KDM2A", account_outliers = FALSE)
}
