% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionickomics_utils.r
\name{get_variants_ClinVar}
\alias{get_variants_ClinVar}
\title{Title}
\usage{
get_variants_ClinVar(gene_of_interest = GOI, pathogenic = TRUE)
}
\arguments{
\item{gene_of_interest}{string of the HGNC accepted gene symbol}

\item{pathogenic}{boolean for limiting to pathogenic variants only}
}
\value{
dataframe
}
\description{
Title
}
\examples{
get_variants_ClinVar("CRB1", pathogenic=TRUE)
}
