% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionickomics_utils.r
\name{import_ClinVar}
\alias{import_ClinVar}
\title{Import ClinVar}
\usage{
import_ClinVar(ClinVarFile, GeneOfInterest, pathogenic = TRUE)
}
\arguments{
\item{pathogenic}{}
}
\value{
dataframe
}
\description{
Imports the downloaded ClinVar data file, filters for the \code{GeneOfInterest}, keeping only known \code{pathogenic} variants. If \code{pathogenic} is set to \code{FALSE} all variants will be returned.
}
