% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cohort.R
\name{import_cohort}
\alias{import_cohort}
\title{Import the cohort data}
\usage{
import_cohort(path_to_cohort, p_code_column)
}
\arguments{
\item{path_to_cohort}{filepath to the cohort data, can be .tsv or .csv}

\item{p_code_column}{the name of the column that contains the variant's p.code (required)}
}
\value{
A tibble containing the cohort data with variant frequency and protein positon added
}
\description{
Function to import cohort data (.csv or .tsv) and adding variant frequency and protein position
}
\examples{
examplepath = system.file("extdata", "MBOAT7_patient_variants.csv", package="LolliplotR")
MBOAT7_cohort = import_cohort(path_to_cohort = examplepath, p_code_column = "Variant_p")
}
