% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionickomics_utils.r
\name{process_HuVarOutput}
\alias{process_HuVarOutput}
\title{Process variants from HuVarBase output}
\usage{
process_HuVarOutput(filename = huvarfile, phenotypes_to_keep = "none")
}
\arguments{
\item{filename}{path to the HuVar output saved data}

\item{phenotypes_to_keep}{list object of characters detailing the disease phenotypes to be kept}
}
\value{
dataframe of cleaned and filtered HuVar single nucleotide variants
}
\description{
\code{process_HuVarOutput} allows the importing of raw HuVarBase variant information to be suitable for input into lolliplotR.
}
\details{
For further information on HuVarBase, check out \href{https://doi.org/10.1371/journal.pone.0210475}{HuVarBase: A human variant database with comprehensive information at gene and protein levels}

\emph{Note}: currently the HuVarBase server is under maintenance -  information can be copied and pasted into text format and used within this function.
}
