% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionickomics_utils.r
\name{rescale_introns}
\alias{rescale_introns}
\title{Rescale introns}
\usage{
rescale_introns(grange, minIntronSize)
}
\arguments{
\item{grange}{\code{GRanges} object of the transcript model for the gene of interest}

\item{minIntronSize}{\code{numeric} vector to fix all \code{introns} to for representation}
}
\value{
rescaled \code{GRanges} object with updated \code{intron} sizes
}
\description{
Rescale introns
}
\examples{
rescale_introns(GRangeObject, IntronSize)
}
