% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionickomics_utils.r
\name{rescale_variants}
\alias{rescale_variants}
\title{Rescale variants based upon new intronic sizes (fixed)}
\usage{
rescale_variants(variantInfo, transcript_grange, minIntronSize)
}
\arguments{
\item{variantInfo}{`dataframe` of variant information for gene of interest}

\item{transcript_grange}{`GRanges` object of the transcript model for the gene of interest}

\item{minIntronSize}{`numeric` vector to fix all `introns` to for representation}
}
\description{
Rescale variants based upon new intronic sizes (fixed)
}
