% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionickomics_utils.r
\name{split_into_multiple}
\alias{split_into_multiple}
\title{Title
https://stackoverflow.com/questions/4350440/split-data-frame-string-column-into-multiple-columns#4351918}
\usage{
split_into_multiple(column, pattern = ", ", into_prefix)
}
\arguments{
\item{into_prefix}{}
}
\description{
Title
https://stackoverflow.com/questions/4350440/split-data-frame-string-column-into-multiple-columns#4351918
}
